pyproject.toml
[project]
name = "caper"
dynamic = ["version"]
description = "Cromwell Assisted Pipeline ExecutoR"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Jin Lee", email = "leepc12@gmail.com" },
    { name = "Mihir Samdarshi", email = "msamdars@stanford.edu"}
]
maintainers = [
    { name = "Mihir Samdarshi", email = "msamdars@stanford.edu"}
]
keywords = ["cromwell", "pipeline", "workflow", "bioinformatics"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "pyhocon>=0.3.53",
    "requests>=2.20",
    "pyopenssl",
    "autouri>=0.4.4",
    "miniwdl>=0.7.0",
    "humanfriendly",
    "numpy>=1.8.2",
    "pandas>=1.0",
    "scikit-learn>=0.19.2",
    "matplotlib>=1.5",
    "six>=1.13.0",
]

[project.urls]
Homepage = "https://github.com/MoTrPAC/caper"
Repository = "https://github.com/MoTrPAC/caper"

[project.scripts]
caper = "caper.cli:main"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["caper"]

[tool.setuptools.package-data]
caper = ["bin/*", "scripts/*"]

[tool.setuptools.dynamic]
version = { attr = "caper.__version__" }

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff",
    "pyrefly",
]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
markers = [
    "integration: Integration tests",
    "google_cloud: Tests using Google Cloud services (require auth)",
    "slow: mark test as slow"
]

[tool.ruff]
line-length = 100
show-fixes = true
target-version = "py312"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",
    "ANN401",
    "D203",
    "D212",
    "D401",
    "PTH",
]
# Specify per-file ignores for Ruff
[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D100", "D102", "D103", "S101", "I001", "SLF001", "ANN"]
# CLI module uses arg() pattern designed for dataclass defaults
"caper/cli/args/*.py" = ["RUF009"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.pyright]
pythonVersion = "3.12"
